fix: constant overflow when compiling for 32bit machines (#318)
on Go 1.x(x) all constants are ints by default which causing
build for 32bit machines to fail due to overflow
https://github.com/golang/go/issues/23086

+++ nodedb.go
@@ -230,7 +230,7 @@ func (ndb *nodeDB) DeleteVersionsFrom(version int64) error {
 	})
 
 	// Finally, delete the version root entries
-	ndb.traverseRange(rootKeyFormat.Key(version), rootKeyFormat.Key(math.MaxInt64), func(k, v []byte) {
+	ndb.traverseRange(rootKeyFormat.Key(version), rootKeyFormat.Key(int64(math.MaxInt64)), func(k, v []byte) {
 		ndb.batch.Delete(k)
 	})
